refactor: update error handling to use centralized constants
- Add explicit null and undefined checks across all validation code
- Update error messages to use constants from constants.ts
- Fix test assertions to match new error message formats
- Apply modern TypeScript best practice: check both null and undefined explicitly

All 111 tests passing.
