+++ ssh-control-mcp/src/utils.ts
@@ -9,7 +9,7 @@ import { resolve } from 'path';
  * @throws An error if the file path is null or undefined
  */
 export function expandTilde(filePath: string): string {
-  if (filePath === undefined || filePath === null) {
+  if (filePath === null || filePath === undefined) {
     throw new Error(NULL_OR_UNDEFINED_ARGUMENTS_ERROR);
   }
   if (filePath === '~') {
